refactor(FinishBoardView): FinishBoardView 분리
